==> notebook/finetune/0_prepare_csv.ipynb <==
import pandas as pd
import re
from sklearn.model_selection import train_test_split

pwd

common_train = pd.read_csv('backup/csv/common_train.csv', index_col=0)
common_test = pd.read_csv('backup/csv/common_test.csv', index_col=0)
common_val = pd.read_csv('backup/csv/common_val.csv' , index_col=0)

common_train.head(2)

health_df = pd.read_json('backup/Dataset/Healthcare/transcription.json')
smarthome_df =  pd.read_json('backup/Dataset/Smarthome/transcription.json')

health_df['path'] = health_df['filename'].apply(lambda x : 'Dataset/Healthcare/Record/' + x + '.wav')
smarthome_df['path'] = smarthome_df['filename'].apply(lambda x : 'Dataset/Smarthome/Record/' + x + ".wav")

elder_df = pd.concat([health_df, smarthome_df])

elder_df = elder_df.rename(columns= {"transcript":"sentence"})
elder_df = elder_df[["path","sentence"]]

elder_train, temp = train_test_split(elder_df, test_size=0.2, random_state=42)
elder_val, elder_test = train_test_split(temp, test_size=0.5, random_state=42)

elder_val

dataset_train = pd.concat([common_train, elder_train])
dataset_test = pd.concat([common_test, elder_test])
dataset_val = pd.concat([common_val, elder_val])

dataset_train

def clean_data(data):
    pattern = r"[^ก-๙a-zA-Z0-9\s]+"
    cleaned_data = re.sub(pattern, '', data)
    return cleaned_data

dataset_train['sentence'] = dataset_train['sentence'].apply(clean_data)
dataset_test['sentence'] = dataset_test['sentence'].apply(clean_data)
dataset_val['sentence'] = dataset_val['sentence'].apply(clean_data)

dataset_train.reset_index(drop=True, inplace=True)
dataset_test.reset_index(drop=True, inplace=True)
dataset_val.reset_index(drop=True, inplace=True)

dataset_train['path'] = dataset_train['path'].apply(lambda x : 'backup/' + x)
dataset_test['path'] = dataset_test['path'].apply(lambda x : 'backup/' + x)
dataset_val['path'] = dataset_val['path'].apply(lambda x : 'backup/' + x)

dataset_train

# Prepare

dataset_train['file_name'] = dataset_train['path'].apply(lambda x : x.split('/')[-1])
dataset_test['file_name'] = dataset_test['path'].apply(lambda x : x.split('/')[-1])
dataset_val['file_name'] = dataset_val['path'].apply(lambda x : x.split('/')[-1])

dataset_train

dataset_train['destination'] = dataset_train['file_name'].apply(lambda x : 'wavs/' + x)
dataset_test['destination'] = dataset_test['file_name'].apply(lambda x : 'wavs/' + x)
dataset_val['destination'] = dataset_val['file_name'].apply(lambda x : 'wavs/' + x)

dataset_train

def filter_dataframe(df, patterns, column):
    return df[~df[column].str.contains('|'.join(patterns))]


patterns = ['มึง', 'เพื่อไทย', 'ชุมนุม', 'สัส', 'เหี้ย', 'ไอ้', 'รัฐประหาร', 'กษัตริย์', 'ทหาร', '#', 'เซ็กส์', 'ตำรวจ', 
            'แม่ง', 'ลงถนน', 'กล่าว', 'กองทัพภูมิ', 'สถาบัน', 'รัฐ', 'ประชาธิปไตย', 'พระ', 'ฉัน', 'สภา', 'พูด', 'คุณ', 
            'ผม', 'แฟน', 'กรรม', 'เศรษฐกิจ', 'มัน', 'เมีย', 'ประเทศ', 'ตาย', 'ศพ', 'มาตรการ', 'ท้องถิ่น', 'เขา', 'ตำแหน่ง', 
            'ยิง', 'งบประมาณ', 'เคอร์ฟิว', 'สาธารณสุข', 'การเมือง', 'นาง', 'แม่มด', 'สัตว์', 'สงคราม', 'ปืน', 'บรรพบุรุษ', 
            'นายกรัฐมนตรี', 'อวัยวะ', 'กาลกิณี', 'ราษฎร', 'เสื้อแดง', 'ห่วย', 'และ', 'เรา', ' - ', 'กู', 'แผ่นดิน', 'เจ้า', 'อย่าง', 
            'ถาม', 'ราชการ', 'ภาค', 'บุคคล', 'สถานการณ์', 'พื้นที่', 'เธอ', 'ประชาชน', 'เกลียดชัง', 'จำเลย', 'ตุลาการ', 
            'ระบบ', 'พรบ.', 'ศาล', 'สำหรับ', 'ช้าง', 'สิงห์', 'สังคม', 'ข่าว', 'บริษัท', 'โคตร', "\- ", 'นโยบาย', 'ซึ่ง', 
            'การ', 'พรรค', 'บริหาร', "\-", 'ชุมชน', 'รัฐบาล', 'ศักดินา', 'ผู้', 'ราช', 'กฎหมาย', 'ทักษิณ', 'ปัจจุบัน', 
            'ละคร', 'โรงพยาบาล', 'ประเภท', 'เพื่อ', 'คือ', 'นาย', 'ไทย', 'พัฒนา', 'หนังโป๊', 'โป๊', 'ละเมิด', 'มนุษย์', 
            'หรือ', 'เตียง', 'ครู', 'เลว', 'พระ', 'เด็ก', 'โดย', 'เถื่อน', 'หมอยา', 'บุพบท', 'โรค', 'เบียร์', 'พุทธ', 
            'ชาติ', 'มงคล', 'ได้แก่', 'ภาษี', 'ผิว', 'ลักษณะ', 'กระท่อม', 'ศาสนา', 'ธนาคาร', 'สำเร็จ', 'เคอร์', 'รักษา', 
            'ธรรม', 'วิหาร', 'หลั่ง', 'ซีเลียม', 'ดาเลียน', 'อำมาตย์', 'ไพร่', 'วิจารณ์', 'เวที', 'สิทธิ', 'วัคซีน']


dataset_train = filter_dataframe(dataset_train, patterns, 'sentence')
dataset_test = filter_dataframe(dataset_test, patterns, 'sentence')
dataset_val =  filter_dataframe(dataset_val , patterns, 'sentence')

import os
os.makedirs('wavs',exist_ok=True)


import os
import shutil
from concurrent.futures import ThreadPoolExecutor
import pandas as pd
from tqdm import tqdm

def copy_file(row_tuple):
    index, row = row_tuple
    src = row['path']
    dst = row['destination']
    shutil.copyfile(src, dst)


dataframes = [dataset_train, dataset_test, dataset_val]

for df in dataframes:
    with ThreadPoolExecutor() as executor:
        list(tqdm(executor.map(copy_file, df.iterrows()), total=df.shape[0]))


len(os.listdir('/project/lt200007-tspai2/thepeach/wavs'))

dataset_train_mod = dataset_train[['destination', 'sentence']].rename(columns={'destination':'path'})
dataset_test_mod = dataset_test[['destination', 'sentence']].rename(columns={'destination':'path'})
dataset_val_mod = dataset_val[['destination', 'sentence']].rename(columns={'destination':'path'})


dataset_train_mod

os.makedirs('csv',exist_ok=True)

dataset_train_mod.to_csv('csv/train.csv')
dataset_test_mod.to_csv('csv/test.csv')
dataset_val_mod.to_csv('csv/validation.csv')

# Combine with augment data

import pandas as pd

# import pandas as pd

# path_dict = {
#     'train': 'csv/train.csv',
#     'test': 'csv/test.csv',
#     'val': 'csv/validation.csv'
# }

# path_dict_room = {
#     'train': 'csv/room/train.csv',
#     'test': 'csv/room/test.csv',
#     'val': 'csv/room/validation.csv'
# }

# path_dict_exhibition = {
#     'train': 'csv/exhibition/train.csv',
#     'test': 'csv/exhibition/test.csv',
#     'val': 'csv/exhibition/validation.csv'
# }

# dfs = {split: pd.read_csv(path, index_col=0)[["path", "sentence"]] for split, path in path_dict.items()}
# dfs_room = {split: pd.read_csv(path, index_col=0)[["path", "sentence"]] for split, path in path_dict_room.items()}
# dfs_exhibition = {split: pd.read_csv(path, index_col=0)[["path", "sentence"]] for split, path in path_dict_exhibition.items()}


# dfs['train']

# import os

# os.makedirs('csv/concatenated',exist_ok=True)
# # Concatenation
# dfs_concat = {
#     'train': pd.concat([dfs['train'], dfs_room['train'], dfs_exhibition['train']]),
#     'test': pd.concat([dfs['test'], dfs_room['test'], dfs_exhibition['test']]),
#     'val': pd.concat([dfs['val'], dfs_room['val'], dfs_exhibition['val']])
# }

# # Exporting to csv
# for split, df in dfs_concat.items():
#     df.to_csv(f'csv/concatenated/{split}.csv')
